+++ tests/integration/MultipaymentTest.php
@@ -14,8 +14,8 @@ class MultipaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );
 
@@ -36,8 +36,8 @@ class MultipaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );
 
@@ -58,8 +58,8 @@ class MultipaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );