+++ tests/integration/PaymentTest.php
@@ -14,8 +14,8 @@ class PaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );
 
@@ -41,8 +41,8 @@ class PaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );
 
@@ -68,8 +68,8 @@ class PaymentTest extends TestCase
         $accessToken = $envVars['ACCESS_TOKEN_3DS'] ?? null;
         $sdk = new Sdk(
             $accessToken,
-            'BP1EF6QIC4P001KBGQ10',
-            'BC32CANTRPP001U8NHO0',
+            'ppcoreinternal',
+            'ppcoreinternal',
             ''
         );