+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V1.10.0
+### Added
+- `MelidataError` entity to register errors in Melidata using the Core Monitor service API
+- Integration tests for melidataError scenario
+### Changed
+- Change the platform_ids used in integrated tests with the Core P&P platform_id (`ppcoreinternal`)
+
 ## V1.9.1
 ### Changed
 - Changed second credit card informations in `CardToken` and `MultipaymentTest` from "amex" to "visa"